use go modules
Make use of "go mod" to manage 3. party dependencies and don't have to
rely on a gopath.
Because we don't have a GOPATH anymore, some paths needed to be changed
during protoc compilation.
The Makefile now also ensures that when "make test" is done, the current version is tested.
The plugin and validator.pb.go is always rebuild before running the
tests.
Also a "clean" Makefile target was added to remove generated files.

The validator.pb.go file was regenerated with protoc 3.6.1 and the
newest protoc-gen-go and protoc-gen-gogo versions.